Reallocate agent-rt classes during the build
This commit reallocate all agent-rt classes to a new package
generated during the build. A new package name is generated
for each build version. The package name include the first
characters of the git commit. This way, we can easily verify
the version looking at the the fully qualified class names.

The purpose of these changes is: when using the agent-rt we
don't want to instrument the agent's own classes. A simpler
approach is to ignore BA-DUA packages (9f8d44f), but this
come whith the drawback that we can't use the agent to
collect coverage of ignored packages. This commit fix that.

This is the same approach used by JaCoCo.