Catch any runtime exception and re-throw
BA-DUA update the coverage before any xRETURN/ATHROW instructions.
If a method exit abruptly with a RuntimeException the coverage
is never updated.

This commit introduces a catch-all RuntimeException exception handler
for all methods. The handler simply re-throw the exception with ATHROW
instruction. With this simple approach we can update the coverage of
methods that throws implicit RuntimeException exceptions.