Extracted the IRuntime interface and make RT implement it
Following the plan to allow different mechanisms to provide the execution
data array, normally, the mechanism are runtime specific and abstracted by
the interface `IExecutionDataAccessorGenerator`.

Implementations are provided by `IRuntime` interface implementations and
are used by the instrumentation process. This interface represents a
particular mechanism to collect execution information in the target VM
at runtime.