Created module `ba-dua-example`
This module contains only a single class with companion tests.
The tests are executed with BA-DUA agent and after tests execution
we generate XML coverage report and verify its contents are correct.

The idea is show how we can execute BA-DUA with Maven. We know that
current implementation is not production-ready, but will be useful
for testing BA-DUA builds with different JVMs.

*The implementation works as follow:*

We copy the BA-DUA agent using `maven-dependency-plugin` and execute
tests with the agent loaded through option `<argLine>` of
`maven-surefire-plugin`.

After tests execution we generate coverage report using BA-DUA CLI
with help of `exec-maven-plugin`. Finally, we verify that the XML
report is correct by matching the file checksum using
`maven-enforcer-plugin`.