(alternative fix)
The previous commit fixes the bug by changing the class that introduced
the regression, fair enough.

I don't think it is the correct approach. We were adding unnecessary
instruction to fix a defect in the method instrumentation process.

Changing the assumptions presented on commit 0af4a34 will not be the
correct fix?

This commit reverts the previous one and fixes the problem by a
defensive null check on `CoverageMethodTransformer`, so we do not need
to expect all method's basic blocks have at least one bytecode
instruction.