(definitive fix)
The previous commit fixes the bug by changing the class affected by the
regression.

I don't think it is the correct approach either. We change the stable
assumptions, at least for Java programs, and I don't know all the side
effects of the changes. Maybe in the future, I can do a formal analysis
of it.

So I made a question:

How will be the ASM representation of a Java method with a jump arc to
the first instructions and wrapped with a try/catch block?

The answer is that a single label object (or `visitLabel` method call)
is used.

Assume the following code sequence (pseudo-bytecode):

```
label: <----------
  inst 1;        |
  inst 2;        |
  .              |
  .              |
  .              |
  goto label; ----
  .
  .
```

If this code is wrapped by a try/catch block, the following code will be
generated (at least from the ASM point-of-view):

```
label: <----------
  inst 1;        |
  inst 2;        |
  .              |
  .              |
  .              |
  goto label; ----
  .
  .
end:
handle:
 athrow;
```

And the Exception table is as follows:

```
   from     to     target     type
   label    end    handle     java/lang/Throwable
```

This commit reverts the previous one and changes
`CatchAndThrowMethodVisitor` to generate a code precisely as described
above.